+++ test/ring/middleware/http_response_test.clj
@@ -9,15 +9,15 @@
 (def bad! (always (bad-request! "BAD")))
 (def bad-result (bad-request "BAD"))
 
-(fact "catch-response"
+(fact "wrap-http-response"
 
   (facts "without middleware exception is thrown"
     (bad ..request..) => bad-result
     (bad! ..request..) => (throws Exception))
 
   (facts "with middleware exceptions are converted into responses"
-    ((catch-response bad) ..request..) => bad-result
-    ((catch-response bad!) ..request..) => bad-result)
+    ((wrap-http-response bad) ..request..) => bad-result
+    ((wrap-http-response bad!) ..request..) => bad-result)
 
   (fact "only response-exceptions are caugh"
-    ((catch-response (always (throw (RuntimeException.)))) ..request..) => (throws Exception)))
+    ((wrap-http-response (always (throw (RuntimeException.)))) ..request..) => (throws Exception)))